refactor(#518): text components dir
