+++ anni/src/components/common/text/Explanation.tsx
